Refactored File Structure
Moved PayPeriod.cs to Models and MockData.cs to Data
