+++ AdminUI/Models/PayPeriod.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http.Connections;
 
-namespace AdminUI.Data
+namespace AdminUI.Models
 {
     public class PayPeriod
     {
