+++ AdminUI/Views/PayPeriod/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model AdminUI.Data.PayPeriod
+﻿@model AdminUI.Models.PayPeriod
 
 @{
     ViewData["Title"] = "Create Pay Period";
