+++ AdminUI/Views/PayPeriod/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@model AdminUI.Data.PayPeriod
+﻿@model AdminUI.Models.PayPeriod
 
 @{
     ViewData["Title"] = "Delete Pay Period";
