+++ AdminUI/Views/PayPeriod/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model AdminUI.Data.PayPeriod
+﻿@model AdminUI.Models.PayPeriod
 
 @{
     ViewData["Title"] = "Edit Pay Period";
