+++ AdminUI/Views/PayPeriod/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<AdminUI.Data.PayPeriod>
+﻿@model IEnumerable<AdminUI.Models.PayPeriod>
 
 @{
     ViewData["Title"] = "Pay Periods";
